src/core_requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile core_requirements.in
#
aniso8601==9.0.1
    # via flask-restful
click==8.0.1
    # via flask
cobble==0.1.3
    # via mammoth
colorama==0.4.4
    # via click
flask==2.0.1
    # via
    #   -r core_requirements.in
    #   flask-api
    #   flask-restful
flask-api==3.0.post1
    # via -r core_requirements.in
flask-restful==0.3.9
    # via -r core_requirements.in
gunicorn==20.1.0
    # via -r core_requirements.in
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
mammoth==1.4.17
    # via -r core_requirements.in
markupsafe==2.0.1
    # via jinja2
python-decouple==3.4
    # via -r core_requirements.in
pytz==2021.1
    # via flask-restful
six==1.16.0
    # via flask-restful
werkzeug==2.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
